Introduce a new pool property 'dedupmax' to limit the size of the DDT
Track the current size of the DDT, and the average entry size in spa_t

Once the maximum size is reached, new entries will not be deduped
Existing entries are still updated and continue to dedup

Extend ddt_lookup() to return whether it added a new item to the DDT

If a new entry is added, the estimated size of the DDT is incremented
by the current average entry size.

During the next ddt_sync() the actual size of the DDT is determined
and the average entry size is updated.

Signed-off-by: Sean Fagan <sean.fagan@klarasystems.com>
Signed-off-by: Allan Jude <allan@klarasystems.com>